refacor: cleaned up start feature architecture
