+++ SceneIt/Features/Start/StartViewModel.swift
@@ -1,16 +1,17 @@
 import Combine
-import SwiftUI
+import Foundation
  
 class StartViewModel: ObservableObject {
+ 
     @Published var state = StartViewState()
  
-    init() {
+    init(onStart: @escaping (Category) -> Void) {
         state.onSelectCategory = { [weak self] category in
             self?.state.selectedCategory = category
         }
         state.onStart = { [weak self] in
-            guard self?.state.selectedCategory != nil else { return }
-            self?.state.showQuiz = true
+            guard let category = self?.state.selectedCategory else { return }
+            onStart(category)
         }
     }
 }
